+++ GestLab/requetebd5.py
@@ -6,9 +6,7 @@ from hashlib import sha256
 from datetime import datetime, timedelta
 import random
 import string
-import smtplib
-import json
-from email.message import EmailMessage
+
 
 
 cnx = ouvrir_connexion()
@@ -436,25 +434,6 @@ def get_all_info_from_domaine(cnx):
 
 # get_info_demande(cnx)
 
-
-def envoyer_mail(mailreceveur, login, mdp):
-    json_file = open('Gestlab/static/data/configEmail.json')
-    gmail_config = json.load(json_file)
-
-    msg = EmailMessage()
-    msg['Subject'] = 'Test email'
-
-    #email de l'envoyeur
-    msg['From'] = gmail_config["email"]
-
-    #email du receveur
-    msg['To'] = mailreceveur
-    msg.set_content('Voici le login: ' + login  + ' \nVoici le mots de passe temporaire: ' + mdp)
- 
-    with smtplib.SMTP_SSL('smtp.gmail.com', gmail_config["port"]) as smtp:
-        smtp.login(gmail_config["email"], gmail_config["password"])
-        smtp.send_message(msg)
-        print("Mail envoyé")
 
 def get_domaine(cnx):
     try:
