+++ GestLab/templates/base.html
@@ -28,7 +28,7 @@
                 {% if session['utilisateur'][1] == 2 %}
                 <li><a href="{{ url_for('demander') }}">demander</a></li>
                 <li><a href="{{ url_for('inventaire') }}">inventaire</a></li>
-                <li><a href="{{ url_for('commentaire') }}">Signaler des alertes</a></li>
+                <li><a href="{{ url_for('commentaire') }}">envoyer commentaire</a></li>
                 {% endif %}
 
                 <!-- laborantin -->
