+++ GestLab/templates/home.html
@@ -48,7 +48,7 @@
         </a>
         <a href="{{ url_for('commentaire') }}">
             <section class="carte-lien">
-                <h2>SIGNALER DES ALERTES</h2>
+                <h2>ENVOYER UN COMMENTAIRE</h2>
                 <p>aux gestionnaires</p>
                 <p>à l'administration</p>
             </section>
